+++ server/src/user/controller/user.controller.ts
@@ -1,5 +1,5 @@
 import { Controller, Get } from '@nestjs/common';
-import { UserService } from './user.service';
+import { UserService } from '../service/user.service';
 
 @Controller('user')
 export class UserController {
